// src/app/api_codegen/model/updateHolidayDto.ts
/**
 * Invicto API
 * REST API Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TypeHolidayRecurrence } from './typeHolidayRecurrence';

export interface UpdateHolidayDto { 
    id: number;
    name?: string;
    recurrence?: TypeHolidayRecurrence;
    startDate?: Date;
    endDate?: Date;
}